+++ src/Form/Parameter/PresentationTextType.php
@@ -1,25 +1,25 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Parameter;
 
-use App\Entity\Parameter\Title;;
 use Symfony\Component\Form\AbstractType;
+use App\Entity\Parameter\PresentationText;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 
-class TitleType extends AbstractType
+class PresentationTextType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('title')
+        $builder->add('presentationText')
         ;
     }
 
     public function configureOptions(OptionsResolver $resolver)
     {
         $resolver->setDefaults([
-            'data_class' => Title::class,
+            'data_class' => PresentationText::class,
         ]);
     }
 }
